+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import kitchenpos.application.exception.OrderTableNotFoundException;
 import kitchenpos.application.exception.TableGroupNotFoundException;
 import kitchenpos.domain.ordertable.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.domain.ordertable.repository.OrderTableRepository;
-import kitchenpos.repository.TableGroupRepository;
+import kitchenpos.domain.tablegroup.repository.TableGroupRepository;
 import kitchenpos.ui.dto.request.CreateOrderGroupOrderTableRequest;
 import kitchenpos.ui.dto.request.CreateTableGroupRequest;
 import org.springframework.stereotype.Service;
