+++ src/main/java/kitchenpos/ui/dto/response/CreateTableGroupResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.ui.dto.response;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public class CreateTableGroupResponse {
